기물 이동을 구현하고 패키지 위치를 변경했습니다. - `ChessGame`에 기물 이동 기능을 구현했습니다. - `java_chess` 패키지 하위로 모든 패키지, 클래스를 이동했습니다. - 'ChessApplication' 명령 분기 방식에 Enum을 적용했습니다.
