--- src/main/java/chess/pieces/enums/Color.java
@@ -1,5 +1,0 @@
-package chess.pieces.enums;
-
-public enum Color {
-    BLACK, WHITE, NONE
-}
