+++ src/main/java/java_chess/chess/pieces/enums/Color.java
@@ -1,0 +1,5 @@
+package java_chess.chess.pieces.enums;
+
+public enum Color {
+    BLACK, WHITE, NONE
+}
