+++ src/main/java/java_chess/utils/StringUtils.java
@@ -1,0 +1,13 @@
+package java_chess.utils;
+
+public class StringUtils {
+
+    private static final String NEW_LINE = System.lineSeparator();
+
+    public static String appendNewLine(String str) {
+        return str + NEW_LINE;
+    }
+
+    private StringUtils() {
+    }
+}
