--- src/main/java/utils/StringUtils.java
@@ -1,13 +1,0 @@
-package utils;
-
-public class StringUtils {
-
-    private static final String NEW_LINE = System.lineSeparator();
-
-    public static String appendNewLine(String str) {
-        return str + NEW_LINE;
-    }
-
-    private StringUtils() {
-    }
-}
