+++ src/test/java/java_chess/application/ChessViewTest.java
@@ -1,0 +1,32 @@
+package java_chess.application;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java_chess.chess.Board;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class ChessViewTest {
+
+    @Test
+    @DisplayName("체스판을 출력한다.")
+    void printBoard() {
+        var board = new Board();
+        var chessView = new ChessView(board);
+        var expectedBoardString = """
+            ♜♞♝♛♚♝♞♜
+            ♟♟♟♟♟♟♟♟
+            □□□□□□□□
+            □□□□□□□□
+            □□□□□□□□
+            □□□□□□□□
+            ♙♙♙♙♙♙♙♙
+            ♖♘♗♕♔♗♘♖
+            """;
+        board.initialize();
+
+        var actualResult = chessView.printBoard();
+        assertThat(actualResult).isEqualTo(expectedBoardString);
+
+    }
+}
